Programming with Socket.IO v.1
Trying to change the TCPJoin script to work with Socket.IO
