+++ Assets/Packages/WebSocketSharp-netstandard.1.0.1.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4737a5e3f58c19441a66bc1cf6af2823
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
