+++ Assets/Packages/WebSocketSharp-netstandard.1.0.1/WebSocketSharp-netstandard.1.0.1.nupkg.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 588cf34e4aa107b408ec468db51b59c0
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
