+++ Assets/Packages/WebSocketSharp-netstandard.1.0.1/lib.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 25f0af3e11a0f8841a0da80dfdd78dc9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
