+++ Assets/Packages/WebSocketSharp-netstandard.1.0.1/lib/netstandard2.0.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ab152ec0c9b295d4d85964d09ad76165
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
