+++ Assets/Packages/WebSocketSharp-netstandard.1.0.1/lib/netstandard2.0/websocket-sharp.dll
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:339c740207f308d9e86b03a4d45d29f17c52476d1ecda88afa9f607966d226fc
+size 226816
