+++ Assets/Packages/WebSocketSharp-netstandard.1.0.1/lib/netstandard2.0/websocket-sharp.xml.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 92739248b2d59464499ff64b3c60dc49
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
