+++ Assets/packages.config
@@ -1,2 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-<packages />
+<packages>
+  <package id="WebSocketSharp-netstandard" version="1.0.1" />
+</packages>
